factor out search operations
